requirements.txt
#Package requirements for carryout_rotor.py

pyserial
regex
socket
